apps/api/requirements.txt
# FastAPI and web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
asyncpg==0.30.0

# Redis and Celery
redis==5.2.1
celery==5.4.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Code quality
ruff==0.8.3

# Utilities
python-multipart==0.0.18
python-dotenv==1.0.1
